requirements.txt
streamlit
PyPDF2
deepseek
python-dotenv